fixed the transaction for deleting seeds

+++ src/services/db/seeds.ts
@@ -4,7 +4,8 @@ import prisma from 'src/auth/lib/prisma/db-prisma';
 
 import { revalidatePath } from 'next/cache';
 import { paths } from 'src/routes/paths';
-import { getUserSettings, updateUserSettings } from './user-settings';
+import { auth } from 'src/auth/lib/mongodb/auth-mongodb';
+import { getUserSettings } from './user-settings';
 
 export const getSeeds = async () => {
   try {
@@ -26,22 +27,37 @@ export const getSeeds = async () => {
 
 export const deleteSeedsByIds = async (ids: string[]): Promise<any | null> => {
   const numberToBeDeleted = ids.length;
+  const session = await auth();
+  const id = session?.user.id;
+  if (!id) {
+    throw new Error('Access denied.');
+  }
   try {
     const result = await prisma.$transaction(async (tx) => {
       await tx.seedBatches.deleteMany({
         where: { id: { in: ids } },
       });
-      const { planPermissionsUsed } = await getUserSettings({ planPermissionsUsed: true });
-      const updatedSeeds = planPermissionsUsed.seeds - numberToBeDeleted;
-      await updateUserSettings(
-        { planPermissionsUsed: { update: { seeds: updatedSeeds } } },
-        { id: true }
-      );
+      await tx.userSettings.update({
+        where: { id },
+        data: {
+          planPermissionsUsed: {
+            update: {
+              seeds: {
+                decrement: numberToBeDeleted,
+              },
+            },
+          },
+        },
+        select: {
+          id: true,
+          appLogin: false,
+        },
+      });
     });
     revalidatePath(paths.seed.root);
     return result;
   } catch (error) {
     console.log('Unable to delete', error);
-    return null;
+    throw new Error("Couldn't delete seeds");
   }
 };
